+++ src/request/timestamp.cpp
@@ -158,7 +158,9 @@ time_t parse_rfc1123_date_time(const char_t*& data, size_t& remaining)
         return -1;
     }
 
-    return seconds_since_epoch(second_of_day, day, month, year);
+    return seconds_since_epoch(
+        static_cast<uint32_t>(second_of_day), static_cast<uint8_t>(day),
+        static_cast<uint8_t>(month), static_cast<uint32_t>(year));
 }
 
 time_t parse_rfc850_date_time(const char_t*& data, size_t& remaining)
@@ -201,7 +203,9 @@ time_t parse_rfc850_date_time(const char_t*& data, size_t& remaining)
         return -1;
     }
 
-    return seconds_since_epoch(second_of_day, day, month, year);
+    return seconds_since_epoch(
+        static_cast<uint32_t>(second_of_day), static_cast<uint8_t>(day),
+        static_cast<uint8_t>(month), static_cast<uint32_t>(year));
 }
 
 time_t parse_asctime_date_time(const char_t*& data, size_t& remaining)
@@ -226,7 +230,9 @@ time_t parse_asctime_date_time(const char_t*& data, size_t& remaining)
         skip_whitespace(data, remaining);
     }
 
-    return seconds_since_epoch(second_of_day, day, month, year);
+    return seconds_since_epoch(
+        static_cast<uint32_t>(second_of_day), static_cast<uint8_t>(day),
+        static_cast<uint8_t>(month), static_cast<uint32_t>(year));
 }
 
 weekday_value_type parse_weekday(const char_t*& data, size_t& remaining)
